Adding n-Tuples
Tuple types are automatically generated by the compiler, when used.
It is possible to generate classes TupleN for N >= 2. Also Tuple0 with the only value () is possible.
A Tuple1 does not exist, since it would be completely useless.

There are three different ways to write tuples and their types:

    Tuple2<Int, String> a := Tuple2<Int, String>(1,"Xy")
    println(a._1)
    println(a._2)

    Tuple2<Int, String> b := (1,"Xy")
    println(b._1)
    println(b._2)

    (Int, String) c := (1,"Xy")
    println(c._1)
    println(c._2)

All three notations do the same.
The second one introduces a syntactic sugar for tuple values,
the third one also introduces a syntactic sugar for tuple types.