unpacking tuples uses tmp vars, to avoid multiple eval of rhs expressions
The type of the temporary variables depends on an expression, so that a simple kind of type
inference is implemented, which allows to infer the type of a variable declaration from the
return type of an expression. A program of the form

    Int a, Int b := (1,2)

is translated into

    tmp := (1,2)
    Int a := tmp._1
    Int b := tmp._2

The type of tmp depends on the type of the expression (1,2), which is resolved in the resolve
visitor and can not be determined earlier.
The temporary variable is needed, since side effects during tuple construction would be repeated
if the expression was evaluated multiple times. The following program:

    Int a := 1
    Int b := 2

    a,b := (b,a)

should exchange the values of a and b. Without the temporary variable, the expression (b,a)
would be evaluated twice and the result would be wrong:

    Int a := 1
    Int b := 2

    a := (b,a)._1  // a becomes 2
    b := (b,a)._2  // b stays 2

The declaration of the temporary variable owns a reference to the rhs expression of the
assignment, which is used to determine the type of tmp.
With this commit, it could be possible to extend Monty by introducing an Auto keyword for
local variable declarations:

    Auto x := 5

This would always work for local variable declarations which are combined with an assignment.