Moving tuples to the core library
In the prior implementation of tuples the tuple type declarations were added
to the current module, which worked good if the tuples only were used inside
that module. Since tuples are core types, they should be located inside the
core library. This also allows the core library to use tuple types.